Feature/validate password and confirm password

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -13,10 +13,17 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
-import java.util.*;
+import java.util.List;
+import java.util.Date;
+import java.util.Objects;
+import java.util.Base64;
+import java.util.StringTokenizer;
+import java.util.ArrayList;
+
 
 @Controller
 public class ImageController {
@@ -45,11 +52,12 @@ public class ImageController {
     //Also now you need to add the tags of an image in the Model type object
     //Here a list of tags is added in the Model type object
     //this list is then sent to 'images/image.html' file and the tags are displayed
-    @RequestMapping("/images/{title}")
-    public String showImage(@PathVariable("title") String title, Model model) {
-        Image image = imageService.getImageByTitle(title);
+    @RequestMapping("/images/{id}/{image}")
+    public String showImage(@PathVariable("id") Integer id, Model model) {
+        Image image = imageService.getImageById(id);
         model.addAttribute("image", image);
         model.addAttribute("tags", image.getTags());
+        model.addAttribute("comments", image.getComments());
         return "images/image";
     }
 
@@ -92,13 +100,21 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session, RedirectAttributes redirectAttributes) {
         Image image = imageService.getImage(imageId);
-
+        if(!isImgOwnerLoggedIn(image, session)){
+            redirectAttributes.addFlashAttribute("editError", true);
+            return "redirect:/images/" + image.getId() + '/' + image.getTitle();
+        }
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
         return "images/edit";
+    }
+
+    private boolean isImgOwnerLoggedIn(Image image, HttpSession session){
+        User user = (User) session.getAttribute("loggeduser");
+        return Objects.equals(user.getId(), image.getUser().getId());
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -132,7 +148,7 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/" + updatedImage.getId() + '/' + updatedImage.getTitle();
     }
 
 
@@ -140,7 +156,12 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, RedirectAttributes redirectAttributes) {
+        Image image = imageService.getImage(imageId);
+        if(!isImgOwnerLoggedIn(image, session)){
+            redirectAttributes.addFlashAttribute("deleteError", true);
+            return "redirect:/images/" + image.getId() + '/' + image.getTitle();
+        }
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
